# openapi.yaml
openapi: 3.0.3
info:
  title: kaytx Comprehensive API
  description: Complete API documentation for the kaytx platform including all endpoints, schemas, and detailed examples.
  version: 2.1.3
  contact:
    name: kaytx API Team
    email: kaytxoffical@gmail.com
    url: https://kaytx.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kaytx.com/v1
    description: Production environment
  - url: https://staging-api.kaytx.com/v1
    description: Staging environment
  - url: http://localhost:3001/v1
    description: Local development
...
paths:
...
  /api/v1/services/agent-counseling/employee-to-agent:
    post:
      tags:
        - Counseling
      summary: Employee initiates counseling with an agent
      operationId: employeeToAgentCounseling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId, topic, question]
              properties:
                agentId:
                  type: string
                topic:
                  type: string
                question:
                  type: string
                options:
                  type: object
                  properties:
                    type:
                      type: string
                    priority:
                      type: string
                    expectedDeliverable:
                      type: string
                    deadline:
                      type: string
                      format: date-time
                    tags:
                      type: array
                      items:
                        type: string
                    confidentiality:
                      type: string
      responses:
        '200':
          description: Counseling session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/services/agent-counseling/agent-to-employee:
    post:
      tags:
        - Counseling
      summary: Agent initiates counseling with the authenticated employee
      operationId: agentToEmployeeCounseling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId, topic, question]
              properties:
                agentId:
                  type: string
                topic:
                  type: string
                question:
                  type: string
                options:
                  type: object
      responses:
        '200':
          description: Counseling session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/services/agent-counseling/sessions/{sessionId}/respond-as-employee:
    post:
      tags:
        - Counseling
      summary: Employee responds to a counseling session
      operationId: respondToCounselingAsEmployee
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [response]
              properties:
                response:
                  type: object
                  required: [status, answer, confidence, reasoning, requiresFollowUp]
                  properties:
                    status:
                      type: string
                    answer:
                      type: string
                    recommendations:
                      type: array
                      items:
                        type: string
                    confidence:
                      type: number
                    reasoning:
                      type: string
                    caveats:
                      type: array
                      items:
                        type: string
                    requiresFollowUp:
                      type: boolean
                    suggestedNextSteps:
                      type: array
                      items:
                        type: string
      responses:
        '200':
          description: Counseling session updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/services/agent-counseling/sessions/{sessionId}/auto-respond:
    post:
      tags:
        - Counseling
      summary: Auto-generate a counseling response using an AI provider (OpenAI or Gemini)
      operationId: autoRespondToCounseling
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                respondingAgentId:
                  type: string
                provider:
                  type: string
                  enum: [openai, google]
                model:
                  type: string
                temperature:
                  type: number
                maxTokens:
                  type: number
                status:
                  type: string
                  enum: [in_progress, completed]
      responses:
        '200':
          description: Counseling session updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /payments/intents:
    post:
      tags:
        - Payments
      summary: Create payment intent
      operationId: createPaymentIntent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentIntentRequest'
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /payments/methods:
    get:
      tags:
        - Payments
      summary: List payment methods
      operationId: listPaymentMethods
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payment methods retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentMethods:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaymentMethod'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Payments
      summary: Add payment method
      operationId: addPaymentMethod
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPaymentMethodRequest'
      responses:
        '201':
          description: Payment method added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethod'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Marketing endpoints
  /marketing/campaigns:
    get:
      tags:
        - Marketing
      summary: List campaigns
      operationId: listCampaigns
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Campaigns retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaigns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Marketing
      summary: Create campaign
      operationId: createCampaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Analytics endpoints
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get dashboard analytics
      operationId: getDashboardAnalytics
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Dashboard analytics retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardAnalytics'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreatePaymentIntentRequest:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
        currency:
          type: string
          default: USD
        description:
          type: string
        metadata:
          type: object
    PaymentIntent:
      type: object
      properties:
        id:
          type: string
        amount:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [requires_payment_method, requires_confirmation, requires_action, processing, succeeded, canceled]
        clientSecret:
          type: string
    PaymentMethod:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [card, bank_account]
        last4:
          type: string
        brand:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        isDefault:
          type: boolean
    AddPaymentMethodRequest:
      type: object
      required:
        - paymentMethodId
      properties:
        paymentMethodId:
          type: string
        isDefault:
          type: boolean
          default: false

    # Marketing schemas
...
    DashboardAnalytics:
      type: object
      properties:
        overview:
          $ref: '#/components/schemas/OverviewMetrics'
        charts:
          type: array
          items:
            $ref: '#/components/schemas/ChartData'
        reports:
          type: array
          items:
            $ref: '#/components/schemas/Report'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - bearerAuth: []